pyproject.toml
[project]
name = "auto-browser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "black>=25.1.0",
    "google-genai>=1.26.0",
    "html5lib>=1.1",
    "lxml>=6.0.0",
    "openai>=1.97.0",
    "pandas>=2.3.1",
    "playwright>=1.53.0",
    "selenium>=4.34.2",
]
